feat(ui): refactor `SideToggler` component HBL-199 #time 15m
