+++ libs/ui/components/src/lib/SideToggler/index.ts
@@ -1,0 +1,3 @@
+import SideToggler from './SideToggler';
+
+export default SideToggler;
